+++ sdk/identity/identity/src/credentials/managedIdentityCredential/index.ts
@@ -20,6 +20,16 @@ import { fabricMsi } from "./fabricMsi";
 const logger = credentialLogger("ManagedIdentityCredential");
 
 /**
+ * Options to send on the {@link ManagedIdentityCredential} constructor.
+ */
+export interface ManagedIdentityCredentialOptions extends TokenCredentialOptions {
+  /**
+   * Allows specifying a custom resource Id.
+   */
+  resourceId?: string;
+}
+
+/**
  * Attempts authentication using a managed identity available at the deployment environment.
  * This authentication type works in Azure VMs, App Service instances, Azure Functions applications,
  * Azure Kubernetes Services, Azure Service Fabric instances and inside of the Azure Cloud Shell.
@@ -30,6 +40,7 @@ const logger = credentialLogger("ManagedIdentityCredential");
 export class ManagedIdentityCredential implements TokenCredential {
   private identityClient: IdentityClient;
   private clientId: string | undefined;
+  private resourceId: string | undefined;
   private isEndpointUnavailable: boolean | null = null;
 
   /**
@@ -39,27 +50,29 @@ export class ManagedIdentityCredential implements TokenCredential {
    * @param clientId - The client ID of the user-assigned identity, or app registration (when working with AKS pod-identity).
    * @param options - Options for configuring the client which makes the access token request.
    */
-  constructor(clientId: string, options?: TokenCredentialOptions);
+  constructor(clientId: string, options?: ManagedIdentityCredentialOptions);
   /**
    * Creates an instance of ManagedIdentityCredential
    *
    * @param options - Options for configuring the client which makes the access token request.
    */
-  constructor(options?: TokenCredentialOptions);
+  constructor(options?: ManagedIdentityCredentialOptions);
   /**
    * @internal
    * @hidden
    */
   constructor(
-    clientIdOrOptions: string | TokenCredentialOptions | undefined,
-    options?: TokenCredentialOptions
+    clientIdOrOptions: string | ManagedIdentityCredentialOptions | undefined,
+    options?: ManagedIdentityCredentialOptions
   ) {
     if (typeof clientIdOrOptions === "string") {
       // clientId, options constructor
       this.clientId = clientIdOrOptions;
+      this.resourceId = options?.resourceId;
       this.identityClient = new IdentityClient(options);
     } else {
       // options only constructor
+      this.resourceId = clientIdOrOptions?.resourceId;
       this.identityClient = new IdentityClient(clientIdOrOptions);
     }
   }
@@ -68,7 +81,6 @@ export class ManagedIdentityCredential implements TokenCredential {
 
   private async cachedAvailableMSI(
     scopes: string | string[],
-    clientId?: string,
     getTokenOptions?: GetTokenOptions
   ): Promise<MSI> {
     if (this.cachedMSI) {
@@ -78,7 +90,15 @@ export class ManagedIdentityCredential implements TokenCredential {
     const MSIs = [fabricMsi, appServiceMsi2017, cloudShellMsi, arcMsi, tokenExchangeMsi(), imdsMsi];
 
     for (const msi of MSIs) {
-      if (await msi.isAvailable(scopes, this.identityClient, clientId, getTokenOptions)) {
+      if (
+        await msi.isAvailable({
+          scopes,
+          identityClient: this.identityClient,
+          clientId: this.clientId,
+          resourceId: this.resourceId,
+          getTokenOptions,
+        })
+      ) {
         this.cachedMSI = msi;
         return msi;
       }
@@ -91,7 +111,6 @@ export class ManagedIdentityCredential implements TokenCredential {
 
   private async authenticateManagedIdentity(
     scopes: string | string[],
-    clientId?: string,
     getTokenOptions?: GetTokenOptions
   ): Promise<AccessToken | null> {
     const { span, updatedOptions } = createSpan(
@@ -101,13 +120,14 @@ export class ManagedIdentityCredential implements TokenCredential {
 
     try {
       // Determining the available MSI, and avoiding checking for other MSIs while the program is running.
-      const availableMSI = await this.cachedAvailableMSI(scopes, clientId, updatedOptions);
+      const availableMSI = await this.cachedAvailableMSI(scopes, updatedOptions);
 
       return availableMSI.getToken(
         {
           identityClient: this.identityClient,
           scopes,
-          clientId,
+          clientId: this.clientId,
+          resourceId: this.resourceId,
         },
         updatedOptions
       );
@@ -147,7 +167,7 @@ export class ManagedIdentityCredential implements TokenCredential {
       // If it's null, it means we don't yet know whether
       // the endpoint is available and need to check for it.
       if (this.isEndpointUnavailable !== true) {
-        result = await this.authenticateManagedIdentity(scopes, this.clientId, updatedOptions);
+        result = await this.authenticateManagedIdentity(scopes, updatedOptions);
 
         if (result === null) {
           // If authenticateManagedIdentity returns null,